running fs.rmSync throws ENOENT if it's not there

+++ gulp/clean.js
@@ -2,7 +2,9 @@ import fs from 'node:fs';
 import { DIST } from './constants.js';
 
 export default function cleanTask(cb) {
-    fs.rmSync(`${DIST}`, { recursive: true });
+    if (fs.existsSync(DIST)) {
+        fs.rmSync(`${DIST}`, { recursive: true });
+    }
     fs.mkdirSync(`${DIST}`, { recursive: true });
     cb?.();
 }
